Refactor speaker module into the speakers module
