+++ app/src/main/res/navigation/main_nav_graph.xml
@@ -77,9 +77,9 @@
     </fragment>
     <fragment
         android:id="@+id/speakerDetailsFragment"
-        android:name="com.android254.droidconKE2020.speaker.views.ui.SpeakerDetailsFragment"
+        android:name="com.android254.droidconKE2020.speakers.ui.views.SpeakerDetailsFragment"
         android:label="Speaker Details"
-        app:moduleName="speaker"
+        app:moduleName="speakers"
         tools:layout="@layout/fragment_speaker_details">
 
         <argument
