+++ buildSrc/local.properties
@@ -4,5 +4,9 @@
 # Location of the SDK. This is only used by Gradle.
 # For customization when using a Version Control System, please read the
 # header note.
+<<<<<<< HEAD
 #Mon Apr 13 02:46:54 EAT 2020
+=======
+#Sat Apr 11 01:32:35 EAT 2020
+>>>>>>> 0e9f7f6... Refactor speaker module into the speakers module
 sdk.dir=/home/oly/Programs/Android/sdk
