--- features/speaker/src/main/java/com/android254/droidconKE2020/speaker/di/di.kt
@@ -1,10 +1,0 @@
-package com.android254.droidconKE2020.speaker.di
-
-import com.android254.droidconKE2020.speaker.views.viewmodels.SpeakerViewModel
-import org.koin.android.viewmodel.dsl.viewModel
-import org.koin.core.module.Module
-import org.koin.dsl.module
-
-val speakerModule: Module = module {
-    viewModel { SpeakerViewModel() }
-}
