--- features/speaker/src/main/java/com/android254/droidconKE2020/speaker/models/SpeakerDetailsModel.kt
@@ -1,10 +1,0 @@
-package com.android254.droidconKE2020.speaker.models
-
-data class SpeakerDetailsModel (
-    val speakerName : String,
-    val twitterHandle : String,
-    val speakerCompany : String,
-    val speakerRole : String,
-    val speakerBio : String,
-    val speakerProfileImg : String
-)
