+++ features/speakers/src/main/AndroidManifest.xml
