+++ features/speakers/src/main/java/com/android254/droidconKE2020/speakers/models/SpeakerDetailsModel.kt
@@ -1,0 +1,10 @@
+package com.android254.droidconKE2020.speakers.models
+
+data class SpeakerDetailsModel (
+    val speakerName : String,
+    val twitterHandle : String,
+    val speakerCompany : String,
+    val speakerRole : String,
+    val speakerBio : String,
+    val speakerProfileImg : String
+)
